refactor: versioning imports and restructure versioning service files
