+++ backend/src/flow44/api/chat.py
@@ -20,12 +20,12 @@ from flow44.api.deps import Permission, ProjectDep, TokenDep, WsProjectDep, WsUs
 from flow44.config import settings
 from flow44.db.chat import ChatRole, get_messages, save_message
 from flow44.db.events import emit_event, get_events, subscribe, unsubscribe
-from flow44.db.heartbeat import clear_heartbeat, is_run_active, touch_heartbeat, try_claim_run
+from flow44.db.heartbeat import clear_heartbeat, touch_heartbeat, try_claim_run
 from flow44.db.pending_plan import delete_pending_plan, get_pending_plan
 from flow44.logic import data_source as ds_logic
 from flow44.sandbox.main import PnpmSandbox
 from flow44.sandbox.manager import sandbox_manager
-from flow44.versioning import service as versioning
+from flow44.services.versioning import service as versioning
 
 logger = logging.getLogger(__name__)
 
@@ -123,32 +123,6 @@ async def _start_agent(project_id: str, sandbox: PnpmSandbox, coro: Any) -> None
     task.add_done_callback(_RUNNING.discard)
 
 
-async def _reject_edit_while_previewing(websocket: WebSocket, sandbox: PnpmSandbox, project_id: str) -> bool:
-    """Refuse a build started while previewing an old version (detached HEAD)."""
-    if await versioning.is_previewing(sandbox, project_id):
-        await websocket.send_json({"type": "error", "message": "Restore this version before editing"})
-        return True
-    return False
-
-
-async def _handle_version_message(
-    websocket: WebSocket, sandbox: PnpmSandbox, project_id: str, msg_type: str, data: dict[str, Any]
-) -> None:
-    if await is_run_active(project_id):
-        await websocket.send_json({"type": "error", "message": "Can't change versions while the AI is working"})
-        return
-    try:
-        if msg_type == "exit_preview":
-            await versioning.exit_preview(sandbox, project_id)
-        elif msg_type == "preview_version":
-            await versioning.preview_version(sandbox, project_id, data["commit_sha"])
-        else:  # restore_version
-            await versioning.restore_version(sandbox, project_id, data["commit_sha"])
-    except Exception:
-        logger.exception("[versioning] %s failed for %s", msg_type, project_id)
-        await websocket.send_json({"type": "error", "message": "Version operation failed"})
-
-
 @http_router.get("/{project_id}/history")
 async def chat_history(project: ProjectDep) -> list[dict[str, Any]]:
     messages = await get_messages(project.id)
@@ -200,7 +174,7 @@ async def chat_ws(  # noqa: C901, PLR0915
 
             try:
                 if msg_type == "message":
-                    if await _reject_edit_while_previewing(websocket, sandbox, project.id):
+                    if await versioning.reject_edit_while_previewing(websocket, sandbox, project.id):
                         continue
 
                     user_content: str = data["content"]
@@ -271,7 +245,7 @@ async def chat_ws(  # noqa: C901, PLR0915
                     state = BuildState.model_validate_json(state_json)
 
                     if action == "accept":
-                        if await _reject_edit_while_previewing(websocket, sandbox, project.id):
+                        if await versioning.reject_edit_while_previewing(websocket, sandbox, project.id):
                             continue
                         await delete_pending_plan(project.id)
                         execute_agent = ExecuteAgent(
@@ -293,7 +267,7 @@ async def chat_ws(  # noqa: C901, PLR0915
                         await _start_agent(project.id, sandbox, plan_agent.rebuild_with_feedback(state, feedback))
 
                 elif msg_type == "fix_error":
-                    if await _reject_edit_while_previewing(websocket, sandbox, project.id):
+                    if await versioning.reject_edit_while_previewing(websocket, sandbox, project.id):
                         continue
 
                     error_message = data.get("error_message", "")
@@ -340,7 +314,7 @@ async def chat_ws(  # noqa: C901, PLR0915
                     )
 
                 elif msg_type in ("preview_version", "restore_version", "exit_preview"):
-                    await _handle_version_message(websocket, sandbox, project.id, msg_type, data)
+                    await versioning.handle_version_message(websocket, sandbox, project.id, msg_type, data)
             except AgentAlreadyRunning:
                 await websocket.send_json(_AGENT_BUSY)
