+++ backend/src/flow44/api/projects.py
@@ -25,7 +25,7 @@ from flow44.db.project_member import list_shared_projects
 from flow44.integrations.s3 import s3_storage
 from flow44.sandbox.idle_reaper import idle_reaper
 from flow44.sandbox.manager import sandbox_manager
-from flow44.versioning import service as versioning
+from flow44.services.versioning import service as versioning
 
 logger = logging.getLogger(__name__)
